LayoutTests/storage/indexeddb/modern/idbtransaction-objectstore-failures-expected.txt

This tests some obvious failures that can happen while calling transaction.objectStore()

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


ALERT: Upgrade needed: Old version - 0 New version - 1
ALERT: [object IDBTransaction] - versionchange
ALERT: [object IDBDatabase]
ALERT: Caught attempt to access empty-named object store on the transaction
ALERT: Caught attempt to access null-named object store on the transaction
ALERT: Caught attempt to access non-existant object store on the transaction
ALERT: Caught attempt to access valid object store on a transaction that is already finishing
ALERT: put failed (because transaction was aborted)
ALERT: First version change transaction unexpected error - [object Event]
PASS successfullyParsed is true

TEST COMPLETE
